+++ core/controller/onetable.php
@@ -54,7 +54,7 @@ class onetableDocLister extends DocLister{
 		if($tpl!=''){
             $this->toPlaceholders(count($this->_docs),1,"display"); // [+display+] - сколько показано на странице.
 
-            $i=0;
+            $i=1;
             $sysPlh=$this->renameKeyArr($this->_plh,$this->getCFGDef("sysKey","dl"));
 			$noneTPL=$this->getCFGDef("noneTPL","");
 			if(count($this->_docs)==0 && $noneTPL!=''){